+++ Code/GraphMol/MolStandardize/Charge.cpp
@@ -264,10 +264,21 @@ std::pair<unsigned int, std::vector<unsigned int>> *Reionizer::weakestIonized(
 }
 
 Uncharger::Uncharger()
-    : pos_h(SmartsToMol("[+,+2,+3,+4;!H0!$(*~[-])]")),
-      pos_noh(SmartsToMol("[+,+2,+3,+4;H0;!$(*~[-])]")),
+    : pos_h(SmartsToMol("[+,+2,+3,+4;!H0;!$(*~[-]),$(*(~[-])~[-])]")),
+      pos_noh(SmartsToMol("[+,+2,+3,+4;H0;!$(*~[-]),$(*(~[-])~[-])]")),
       neg(SmartsToMol("[-!$(*~[+,+2,+3,+4])]")),
-      neg_acid(SmartsToMol("[$([O-][C,P,S]=O),$([n-]1nnnc1),$(n1[n-]nnc1)]")){};
+      neg_acid(SmartsToMol(
+          // carboxylate, carbonate, sulfi(a)te,
+          // and their thio-analogues
+          // (among other less likely structures)
+          "[$([O,S;-][C,S;+0]=[O,S]),"
+          // phosphi(a)te, nitrate
+          // and their thio-analogues
+          "$([O,S;-][N,P;+](=[O,S])[O,S;-]),"
+          // hali(a)te, perhalate
+          "$([O-][Cl,Br,I;+,+2,+3][O-]),"
+          // tetrazole
+          "$([n-]1nnnc1),$([n-]1ncnn1)]")){};
 
 Uncharger::Uncharger(const Uncharger &other) {
   pos_h = other.pos_h;
@@ -277,6 +288,30 @@ Uncharger::Uncharger(const Uncharger &other) {
 };
 
 Uncharger::~Uncharger(){};
+
+void neutralizeNeg(Atom *atom, int hDelta = 1) {
+  atom->setNumExplicitHs(atom->getTotalNumHs() + hDelta);
+  atom->setNoImplicit(true);
+  atom->setFormalCharge(atom->getFormalCharge() + 1);
+  BOOST_LOG(rdInfoLog) << "Removed negative charge.\n";
+  // since we changed the number of explicit Hs, we need to update the
+  // other valence parameters
+  atom->updatePropertyCache(false);
+}
+
+bool neutralizeNegIfPossible(Atom *atom) {
+  bool is_early_atom = isEarlyAtom(atom->getAtomicNum());
+  bool has_hs = atom->getTotalNumHs();
+  if (is_early_atom && !has_hs) {
+    return false;
+  }
+  int hDelta = (is_early_atom ? -1 : 1);
+  // Add hydrogen to negative atom, increase formal charge
+  // Until quaternary positive == negative total or no more negative
+  // acid
+  neutralizeNeg(atom, hDelta);
+  return true;
+}
 
 ROMol *Uncharger::uncharge(const ROMol &mol) {
   BOOST_LOG(rdInfoLog) << "Running Uncharger\n";
@@ -299,88 +334,90 @@ ROMol *Uncharger::uncharge(const ROMol &mol) {
 
   bool needsNeutralization =
       (q_matched > 0 && (n_matched > 0 || a_matched > 0));
-  std::vector<std::pair<int, int>> a_atoms(a_matches.size());
-  std::vector<std::pair<int, int>> n_atoms(n_matches.size());
   std::vector<unsigned int> atomRanks(omol->getNumAtoms());
   if (df_canonicalOrdering && needsNeutralization) {
     Canon::rankMolAtoms(*omol, atomRanks);
   } else {
     std::iota(atomRanks.begin(), atomRanks.end(), 0);
   }
-  for (unsigned int i = 0; i < n_matches.size(); ++i) {
-    int aidx = n_matches[i][0].second;
-    n_atoms[i] = std::make_pair(atomRanks[aidx], aidx);
-  }
-  for (unsigned int i = 0; i < a_matches.size(); ++i) {
-    int aidx = a_matches[i][0].second;
-    a_atoms[i] = std::make_pair(atomRanks[aidx], aidx);
-  }
+  auto getRankIdxPair = [&atomRanks](const MatchVectType &mv) {
+    int aidx = mv.front().second;
+    return std::make_pair(atomRanks[aidx], aidx);
+  };
+  std::vector<std::pair<int, int>> n_atoms;
+  n_atoms.reserve(n_matches.size());
+  std::transform(n_matches.begin(), n_matches.end(), std::back_inserter(n_atoms),
+    getRankIdxPair);
+  std::vector<std::pair<int, int>> a_atoms;
+  a_atoms.reserve(a_matches.size());
+  std::transform(a_matches.begin(), a_matches.end(), std::back_inserter(a_atoms),
+    getRankIdxPair);
   if (df_canonicalOrdering) {
     std::sort(n_atoms.begin(), n_atoms.end());
     std::sort(a_atoms.begin(), a_atoms.end());
   }
 
+
   // Neutralize negative charges
   if (needsNeutralization) {
     // Surplus negative charges more than non-neutralizable positive charges
     int neg_surplus = n_matched - q_matched;
-    if (n_matched > 0 && neg_surplus > 0) {
+    if (neg_surplus > 0) {
       boost::dynamic_bitset<> nonAcids(omol->getNumAtoms());
       nonAcids.set();
-      for (const auto &pr : a_atoms) {
-        nonAcids.reset(pr.second);
+      for (const auto &pair : a_atoms) {
+        nonAcids.reset(pair.second);
       }
-      unsigned int midx = 0;
       // zwitterion with more negative charges than quaternary positive centres
-      while (neg_surplus > 0 && midx < n_atoms.size()) {
-        unsigned int idx = n_atoms[midx++].second;
+      for (const auto &pair : n_atoms) {
+        unsigned int idx = pair.second;
         if (!nonAcids[idx]) {
           continue;
         }
         Atom *atom = omol->getAtomWithIdx(idx);
-
-        if (!isEarlyAtom(atom->getAtomicNum())) {
-          // Add hydrogen to negative atom, increase formal charge
-          // Until quaternary positive == negative total or no more negative
-          // acid
-          atom->setNumExplicitHs(atom->getTotalNumHs() + 1);
-          atom->setNoImplicit(true);
-          atom->setFormalCharge(atom->getFormalCharge() + 1);
-          --neg_surplus;
-          BOOST_LOG(rdInfoLog) << "Removed negative charge.\n";
-        } else if (atom->getTotalNumHs()) {
-          atom->setNumExplicitHs(atom->getTotalNumHs() - 1);
-          atom->setNoImplicit(true);
-          atom->setFormalCharge(atom->getFormalCharge() + 1);
-          --neg_surplus;
-          BOOST_LOG(rdInfoLog) << "Removed negative charge.\n";
+        if (neutralizeNegIfPossible(atom) && !--neg_surplus) {
+          break;
         }
-        // since we changed the number of explicit Hs, we need to update the
-        // other valence parameters
-        atom->updatePropertyCache(false);
       }
     }
 
     // now do the other negative groups if we still have charges left:
-    if (a_matched > 0 && neg_surplus > 0) {
-      unsigned int midx = 0;
+    neg_surplus = a_matched - q_matched;
+    if (neg_surplus > 0) {
+      boost::dynamic_bitset<> skipChargeSep(omol->getNumAtoms());
+      for (const auto &pair : n_atoms) {
+        unsigned int idx = pair.second;
+        Atom *atom = omol->getAtomWithIdx(idx);
+        for (const auto &nbri :
+             boost::make_iterator_range(omol->getAtomNeighbors(atom))) {
+          const auto &nbr = (*omol)[nbri];
+          auto nbrIdx = nbr->getIdx();
+          // if the neighbor has a positive charge,
+          // neutralize only once (e.g., NO3-)
+          if (nbr->getFormalCharge() > 0) {
+            if (!skipChargeSep.test(nbrIdx)) {
+              skipChargeSep.set(nbrIdx);
+            } else {
+              skipChargeSep.set(idx);
+            }
+            break;
+          }
+        }
+      }
       // zwitterion with more negative charges than quaternary positive centres
-      while (neg_surplus > 0 && midx < a_atoms.size()) {
+      for (const auto &pair : a_atoms) {
         // Add hydrogen to first negative acidic atom, increase formal charge
         // Until quaternary positive == negative total or no more negative atoms
-        Atom *atom = omol->getAtomWithIdx(a_atoms[midx++].second);
-        // skip ahead if we already neutralized this
-        if (atom->getFormalCharge() >= 0) {
+        unsigned int idx = pair.second;
+        Atom *atom = omol->getAtomWithIdx(idx);
+        // skip ahead if we already neutralized this or if it is part of a zwitterion
+        if (atom->getFormalCharge() >= 0 || skipChargeSep.test(idx)) {
           continue;
         }
-        atom->setNumExplicitHs(atom->getTotalNumHs() + 1);
-        atom->setNoImplicit(true);
-        atom->setFormalCharge(atom->getFormalCharge() + 1);
-        --neg_surplus;
-        BOOST_LOG(rdInfoLog) << "Removed negative charge.\n";
-        // since we changed the number of explicit Hs, we need to update the
-        // other valence parameters
-        atom->updatePropertyCache(false);
+        neutralizeNeg(atom);
+        if (!--neg_surplus) {
+          break;
+        }
       }
     }
 
@@ -388,26 +425,7 @@ ROMol *Uncharger::uncharge(const ROMol &mol) {
     for (const auto &pair : n_atoms) {
       auto idx = pair.second;
       Atom *atom = omol->getAtomWithIdx(idx);
-      if (!isEarlyAtom(atom->getAtomicNum())) {
-        atom->setNumExplicitHs(atom->getTotalNumHs());
-        atom->setNoImplicit(true);
-        while (atom->getFormalCharge() < 0) {
-          atom->setNumExplicitHs(atom->getNumExplicitHs() + 1);
-          atom->setFormalCharge(atom->getFormalCharge() + 1);
-          BOOST_LOG(rdInfoLog) << "Removed negative charge.\n";
-        }
-      } else if (atom->getTotalNumHs()) {
-        atom->setNumExplicitHs(atom->getTotalNumHs());
-        atom->setNoImplicit(true);
-        while (atom->getFormalCharge() < 0 && atom->getNumExplicitHs()) {
-          atom->setNumExplicitHs(atom->getTotalNumHs() - 1);
-          atom->setFormalCharge(atom->getFormalCharge() + 1);
-          BOOST_LOG(rdInfoLog) << "Removed negative charge.\n";
-        }
-      }
-      // since we chnaged the number of explicit Hs, we need to update the
-      // other valence parameters
-      atom->updatePropertyCache(false);
+      neutralizeNegIfPossible(atom);
     }
   }
 
@@ -447,10 +465,10 @@ ROMol *Uncharger::uncharge(const ROMol &mol) {
           atom->setNumExplicitHs(atom->getNumExplicitHs() + 1);
         }
         BOOST_LOG(rdInfoLog) << "Removed positive charge.\n";
+        // since we changed the number of explicit Hs, we need to update the
+        // other valence parameters
+        atom->updatePropertyCache(false);
       }
-      // since we changed the number of explicit Hs, we need to update the
-      // other valence parameters
-      atom->updatePropertyCache(false);
       if (!netCharge) {
         break;
       }